+++ src/main/java/com/fiap/tc/adapters/driven/infrastructure/outputs/CategoryOutputAdapter.java
@@ -14,7 +14,7 @@ import org.springframework.stereotype.Service;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-import static com.fiap.tc.adapters.driven.infrastructure.persistence.mappers.base.MapperConstants.CATEGORY_MAPPER;
+import static com.fiap.tc.adapters.driven.infrastructure.mappers.base.MapperConstants.CATEGORY_MAPPER;
 import static java.lang.String.format;
 
 @Service
