+++ src/main/java/com/fiap/tc/adapters/driven/infrastructure/outputs/OrderOutputAdapter.java
@@ -1,7 +1,7 @@
 package com.fiap.tc.adapters.driven.infrastructure.outputs;
 
 import com.fiap.tc.adapters.driven.infrastructure.persistence.builders.OrderHistoricBuilder;
-import com.fiap.tc.adapters.driven.infrastructure.persistence.mappers.base.MapperConstants;
+import com.fiap.tc.adapters.driven.infrastructure.mappers.base.MapperConstants;
 import com.fiap.tc.adapters.driven.infrastructure.persistence.repositories.OrderRepository;
 import com.fiap.tc.core.application.ports.out.order.ListOrdersReadyPreparingOutputPort;
 import com.fiap.tc.core.application.ports.out.order.LoadOrderOutputPort;
