+++ src/main/java/com/fiap/tc/adapters/driver/presentation/mappers/ProductImageRequestMapper.java
@@ -1,6 +1,6 @@
 package com.fiap.tc.adapters.driver.presentation.mappers;
 
-import com.fiap.tc.adapters.driven.infrastructure.persistence.mappers.base.MapperEntity;
+import com.fiap.tc.adapters.driven.infrastructure.mappers.base.MapperEntity;
 import com.fiap.tc.adapters.driver.presentation.requests.ProductImageRequest;
 import com.fiap.tc.core.domain.entities.ProductImage;
 import org.mapstruct.Mapper;
